requirements.txt
streamlit
langchain
python-dotenv
langchain_community
faiss-cpu
langchain-openai
beautifulsoup4